+++ mlxtend/data/iris.py
@@ -36,6 +36,19 @@ def iris_data():
         petal length, and petal width.
         y is a 1-dimensional array of the class labels {0, 1, 2}
 
+
+    Note
+    --------
+    The Iris dataset (originally collected by Edgar Anderson) and
+    available in UCI's machine learning repository is different from
+    the Iris dataset available in R(and the one in the original paper
+    by R.A. Fisher [1]). Precisely, there are two data points(row number
+    34 and 37) in UCI's Machine Learning repository are different from the
+    Iris dataset in R (and the one in the original Fisher paper).
+
+    [1] . A. Fisher (1936). "The use of multiple measurements in taxonomic
+    problems". Annals of Eugenics. 7 (2): 179–188
+
     Examples
     -----------
     For usage examples, please see